+++ CHANGES.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Drop Sidekiq < 7 support
+- Remove deprecated `Sidekiq::Throttled.setup!`
 
 ## [1.4.0] - 2024-04-07